Fix `MontConfig`
